refactor: move business logic to hook
