+++ src/modules/boards/create-board/components/create-board.tsx
@@ -1,44 +1,18 @@
-import { useCallback } from "react";
-import { useBoardSelector } from "@/modules/boards/board-selector";
-import {
-  ManageBoardForm,
-  useManageBoardForm,
-  type TManageBoardForm,
-} from "@/modules/boards/manage-board";
-import { usePostCreateBoard } from "@/modules/core.fetching-hooks";
+import { ManageBoardForm } from "@/modules/boards/manage-board";
+import { useCreateBoard } from "../hooks";
 
 export interface CreateBoardProps {
   onClose: () => void;
 }
 
 export function CreateBoard({ onClose }: CreateBoardProps) {
-  const { setId } = useBoardSelector();
-
-  const form = useManageBoardForm({
-    defaultValues: {
-      name: "",
-    },
-  });
-
-  const createBoard = usePostCreateBoard();
-
-  const onSubmit = useCallback(
-    (data: TManageBoardForm) => {
-      createBoard.mutate(data, {
-        onSuccess: (data) => {
-          onClose();
-          setId(data.id);
-        },
-      });
-    },
-    [createBoard, onClose, setId],
-  );
+  const { form, isPending, onSubmit } = useCreateBoard({ onClose });
 
   return (
     <ManageBoardForm
       action="Create"
       form={form}
-      isPending={createBoard.isPending}
+      isPending={isPending}
       onSubmit={onSubmit}
     />
   );
