+++ src/modules/boards/create-board/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-create-board";
